Fixed issues #1 and #2
See:
https://github.com/JLimperg/SalSSuite/issues/1
https://github.com/JLimperg/SalSSuite/issues/2

The were UI problems with the CitizenModulePanel which were
fixed with this commit. Multi-selection was enabled and the table
is not updated automatically any more when deleting citizens.
When a citizen is added, the table now automatically scrolls
down to the newly created row.

Furthermore, the UI was changed to make it scale better when
the enclosing frame's size is changed.