Introducing the HelpBrowser
The HelpBrowser is a utility meant to serve as a connection
between the SalSSuite and its Wiki on Github
(https://github.com/JLimperg/SalSSuite/wiki). It parses
the Markdown files that serve as Wiki pages and
displays them using an EditorPane.

This commit sets up the HelpBrowser by
a) adding its source code;
b) adding the "MarkdownJ" library that is used for Markdown parsing;
c) modifying the ant buildfile so that every jar file will include
   said library.