Fixed issue #11
See
https://github.com/JLimperg/SalSSuite/issues/11

In all tables, TableModels are not replaced any more when
updating the TableModel. Moreover, they now use typed columns
instead of the 'Object' type (f.ex. "Integer" for "ID" columns),
which should increase performance. Therefore the column resize
that was mentioned in the issue occurs less often (although
it could not be eliminated alltogether).