Fixed issue #13 + minor bugfix in CompanyEditingDialog
See
https://github.com/JLimperg/SalSSuite/issues/13

Commas in user input are now replaced by dots before calling
Double.parseDouble(input) when floating point figures are parsed.

By the way also fixed a bug in CompanyEditingDialog where
bad values in the "ID" and "salary" fields for the founder
would result in a failure of editing/adding a company but
no error message would be displayed.