Fixed issue #16 + bugfix in CompanyEditingDialog
See
https://github.com/JLimperg/SalSSuite/issues/16

In the CompanyEditingDialog, if the user attempts to assign
a citizen to the company they are currently creating/editing and
this citizen is already employed at another company, a warning
message is displayed now.

Moreover, fixed a bug where deleting a citizen from the company's
employee list in the CompanyEditingDialog would result in them
being removed from their current company (companyId = -1) although
the user had not yet chosen to commit the changes made in the
dialog by pressing the "approve" button.