Buildfile does not use temporary directory any more
The temporary directory containing all the classes and libraries
was stupid anyway, so I rewrote the buildfile to load the classes
and libraries directly from their respective folders.

Note that the purpose of the 'clean' target has changed: It does
not delete the temporary directory any more, but rather all
output of every target. The 'cleanall' target has therefore
become obsolete and was removed.