Improved company list export
The CompanyModulePanel now orders companies by ID when generating
the company list. Additionally, the list's layout was changed
in order to provide more relevant information for the user,
therefore the 'productDescription' was added to the list and the
total number of jobs was replaced by the number of free jobs.