Minor performance tweak in MagazineClientOrdersPanel
The orders section of the MagazineClient actually already was quite
optimised as far as performance is concerned. Only a small change
was necessary to prevent the MagazineClientOrdersPanel from
reloading all the data whenever a new order is added. However,
this small change should be quite significant when working with
the client, as reloading all the orders data is incredibly
expensive traffic-wise.