Fixed issues #29 & #21 + bugfix in CreditClient
See
https://github.com/JLimperg/SalSSuite/issues/29
https://github.com/JLimperg/SalSSuite/issues/21

Fixed a bug where table rows would not be properly converted
into TableModel rows, leading to very strange behaviour when
using the tables' AutoRowSorter feature.

Moreover fixed a bug in the CreditClient where changing the
endDay or startDay would not result in the total amount being
updated accordingly and fixed another bug where the client
would not check if endDay is earlier than startDay when
editing credits.

By the way also renamed some methods in the AccountManagingPanel.

Sorry for a bit of confusion due to not logically
grouping all those changes...