Addressed issues #25 and #26 + performance improvements
See
https://github.com/JLimperg/SalSSuite/issues/25

All operations that need a lot of time are now executed by
SwingWorkers so that they run in background and do not block the
GUI. Most notably, this affects (re)loading data from the
database over a slow network connection and exporting various
lists, tables etc. to files.

These SwingWorkers are accompanied by ProgressMonitors to give
the user visual feedback about the approximate duration of the
operation. There is some additional work to do here in order
to compute the number of rows in a ResultSet that comes from
a FilterPanel correctly. See the comments in the code.

To automate the process of fetching data from the database and
building a TableModel accordingly, the new abstract class
TableModelUpdater was created.

By the way, the CompanyModulePanel's and the AccountingClient's
performance was improved by turning generic "fetch everything
from the database" queries into more specific ones. Because I
am short of time, there currently is no mechanism to avoid
concurrent modifications of companies or accounting entries.
The relevant points in the code are marked with TODO comments.
This should be fixed as soon as possible.

Moreover, the EmploymentClient now uses a TableModel with defined
column datatypes and does not create a new model whenever the data
is fetched from the database.