Extended the NameToIDClient and fixed issue #18
See
https://github.com/JLimperg/SalSSuite/issues/18

The NameToIDClient was given more relevance by adding a
companies section where the ID, name and product description
of a company can be converted into each other. Therefore
the client was renamed and is now called 'Converter'.

Moreover, it is now possible to omit either the forename
or the surname of a citizen. This will make sense only in
seldom cases, but after all it was an unnecessary restriction
to require both.

In order to make sure the user understands the functionality
of the Converter (especially regarding how it treats text
input), a 'help' tab was added that describes the Converter's
possibilities.

The ant buildfile was extended so that it will now jar the
Converter as well. In its current incarnation, it is quite useful
for a lot of clients, therefore all users should have access
to it. For that purpose also fixed a bug where the default
close operation used to be DISPOSE_ON_CLOSE for standalone
Converters.

By the way, the javadoc and comments were improved and the code
was restructured a bit.