Improved DutyModule and DutyModulePanel a lot
a) Fixed issue #34

See
https://github.com/JLimperg/SalSSuite/issues/34

The duty export functions now export data for all days instead
of only for the current day, and they do this in a practical
format. Instead of the ID, they now export form and name of
citizens, and the tables are now ordered by form, surname and
forename.

b) Moved the export functions from DutyModule to DutyModulePanel

The output mechanism is a bit cumbersome, I think, therefore
the two duty export functions were moved to the DutyModulePanel.
They are now accessible via buttons just as the other list/table
export functions.

The Constants fields which had to do with DutyModule output
therefore became obsolete and were removed. The same applies
to the server's field attendanceTime that used to be part
of the project settings.

c) More responsive GUI for DutyModulePanel

As with the other ModulePanels and Clients, the TableModel
updating was delegated to a SwingWorker, thereby providing a
much more responsive GUI. The same applies to the two
list export methods.