Added an interest model and fixed 2 bugs in CreditClient
a) Added another interest model that the user can choose.
Before this commit, the client would compute interest for
every day the credit was not paid back. Now the user
may choose to instead compute interest based on a fixed
rate for the whole credit.

b) Fixed a bug where adding a new credit would not lead to
a new table row being created as well.

c) Fixed a bug where canceling the dialog when adding a credit
would lead to an SQLException.