Converter bugfix + usability enhancement
During the first day of my own SalS project, it has turned out
that the Converter is about the most often used of all clients.
As it is so useful in numerous situations, all clients (but
the AccountingClient) now display it at startup.

Moreover, the Converter used a deprecated module description
string for authentication that did not correspond to any module
of the current SalSSuite. This leaded to the uncomfortable situation
that only user admin was able to start a Converter, which has
been fixed. The Converter now uses the AdminModule's authentication
string (AdminModule.NAME) in order to determine whether a user may
access it.