Hotfixes for DutyClient
The DutyClient will now ignore login/logout events from dates
other than the current date when determining whether a citizen
is logged in or out. This means that a citizen who has logged
in on July 14, 2011 but has not logged out on that day any more
is not logged out on the next day but logged in, just as the duty
officers will probably expect it.

Moreover, the client now orders log events by time when
determining whether a citizen is logged in or out. This should
usually happen automatically as rows are selected in the order
in which they have been inserted, but there is no guarantee for that
behaviour.

For improved usability, the feedback display now turns blue when
a citizen is logged out and green when they are logged in.