Wiki doc migration
All the files from the Github wiki were moved to the folder
for version 1.0.x documentation. Relative links were modified
and 'ä/ö/ü' as well as 'ß' and '€' were masked using html
entities. A YAML front matter was added to each file indicating
the SalSSuite version the file documents, the layout to use,
and the page title.

A template for documentation was created which is more or less
equal to the default layout for now.